Refactor lib.rs into separate modules
The public API was divided into four separate modules:

- path
- node
- ext
- error

As it was getting a bit lengthy in lib.rs
