+++ src/error.rs
@@ -1,0 +1,21 @@
+use std::ops::Deref;
+
+use nom::error::{convert_error, VerboseError};
+
+/// A JSONPath error
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("invalid JSONPath string:\n{message}")]
+    InvalidJsonPathString { message: String },
+}
+
+impl<T> From<(T, VerboseError<T>)> for Error
+where
+    T: Deref<Target = str>,
+{
+    fn from((i, e): (T, VerboseError<T>)) -> Self {
+        Self::InvalidJsonPathString {
+            message: convert_error(i, e),
+        }
+    }
+}
